Initial impl as a single-file script
The description below is from the original commit, in which the script
was being added to a larger internal+private `docs` repo. In that
context, this script was replacing a clunkier Docs-as-Data rendering pipeline.

Since then, this tool has been extracted from that `docs` repo into this
new standalone repo.

I’m retaining the original commit message below for context and
posterity.

----

This new pipeline is much simpler than the old pipeline:

* Old:
  * A directory named “jekyll” containing a Jekyll site
  * Symlinks bringing the data and templates into the Jekyll site
  * A shell script that invoked `jekyll build` to create HTML files and
    `prettier` to clean up (reformat) the source of those files, and
    rename them *back* to `.md` so that GitHub would render them.
    * Jekyll would first use Liquid to render the dynamic content in the
      Markdown templates, then use a Markdown renderer (Kramdown, I
      think) to render the results into HTML files
* New:
  * A Clojure script

In addition to the “pipeline” being simpler, the cognitive load for
template authors is also lower:

* Old:
  * Markdown templates get rendered into HTML files via Kramdown, but
    then renamed to `.md` so that GitHub will render them. Authors need
    to have an understanding of Liquid, Kramdown, and GitHub’s Markdown
    renderer (CommonMarker).
* New:
  * Markdown templates get rendered into Markdown files. Authors need
    to have an understanding of Selmer and GitHub’s Markdown renderer
    (CommonMarker).

BTW, normally I don’t like including extensions in the filenames of
scripts (`.clj`) but in this case I’ve done so so that my editor will
apply useful syntax highlighting to the script source. Silly, perhaps,
but DX matters.