Prefer MD to HTML; cleanup whitespace
I was reviewing a template and noticed that I was generating a lot more
HTML than necessary. When writing Markdown, it’s best to default to
Markdown, so I replaced a bunch of the unnecessary HTML with Markdown.

I continue to struggle with whitespace issues so I came with with a
hack: in the templates I can include 🆇 at the end of a line that would
be otherwise generated as blank (because it contains only a template
tag like e.g. {% if ... %} or {% for ... %}. After the template is
rendered by Selmer, that line will consist of only whitespace and then
that 🆇. So then the reduce-whitespace function will just remove that
line entirely. It’s a hack, but it works.

That hack is actually quite similar to the whitespace control feature
under consideration to be added to Selmer:
  https://github.com/yogthos/Selmer/issues/170#issuecomment-594778666