Change how db is provided to docs
We recently ported to a new, homegrown, Docs-as-data build system
that loads a database from YAML files, makes that data available to
templates, and then renders the templates into documents.

That system replaced a jury-rigged pipeline that used Jekyll to do
roughly the same thing. And because Jekyll made the datasets available
to template authors by filename, I did the same thing in the new build
system, because why  not. I didn’t give it much thought.

To illustrate, with Jekyll we might have a directory named `data`
containing `technologies.yaml` containing a mapping with the key
`technologies` and `systems/us.yaml` containing a mapping with the key
`systems`. Jekyll would make the data in those files accessible to
templates as `site.data.technologies.technologies` and
`site.data.systems.us.systems`.

With the new system we implemented with #XXX, those datasets would be
accessible to templates using the same names, except without the prefix
`site.data`.

In other words, almost identical, just with a little less unnecessary
boilerplate namespacing.

It seemed reasonable at the time.

Well, it didn’t take long for a major drawback to emerge: with this
scheme it’s not possible to split a large dataset across multiple files.
Or, put another way: splitting a dataset across multiple files would
fragment the dataset into multiple datasets that would be difficult at
best to join within the template, if that’d even be possible.

Take, for instance, the example above. In addition to `systems/us.yaml`
we might also have `systems/uk.yaml`. Splitting the broader dataset —
all our systems — into two files makes it easier to work with the
dataset; the files are smaller and the records are grouped into files
by some convenient facet. But now a template author has no way to
operate on, or iterate over, the total dataset — it is no longer a
single dataset that can be referred to by a single name, but rather
multiple datasets.

If that seems like no big deal, consider that I’ve recently added a new
dataset `repositories` that is a total of ~17300 lines of YAML. I’d like
to create a Repositories document template, that would list all of our
repositories, but that would require that the entire dataset is stored
as a single giant YAML file — which is not great, in terms of
ergonomics. Because of the change in this commit, I’ve been able to
*physically* split the dataset into/across multiple smaller YAML files,
without *logically* splitting the dataset at runtime, i.e. as it is
provided/presented to templates and thereby template authors.

Hence this change.

It’s fairly straightforward: file paths are no longer included in the
keys of the “context map” (aka the root data value) that’s provided to
the templates at render time. Instead, the root values of each data file
must be a mapping, and all the mappings are merged into a single map
that contains the entire database in a single collection.

The merge that’s used is a deep merge, so a dataset can be authored and
stored in multiple YAML files; as long as they use the same root key,
the values in those files will be merged into a single collection.

In our example above, whereas the *systems* dataset used to be broken
into two datasets: `systems.uk.systems` and `systems.us.systems` it is
now accessible at the simple coordinate `systems` because both files
include the root key `systems`.

This resolves the core problem that we’d really like to be able to
split datasets physically (in files) without splitting them logically
(at runtime).

This also yields a minor ergonomics improvement in terms of the names
used to refer to datasets within templates. Consider a dataset named
`technologies` that is naturally in a file named `technologies.yaml`.
Before now, a template author would have to refer to this dataset as
`technologies.technologies` which is awkward and impairs readability
and therefore maintainability of the templates.

I feel really good about this, it feels right.