Add a new CLI command: `rebuild`
All this does is read the database and write it right back out,
normalizing the formatting.

Super simple, but I’ve found myself firing up a REPL a few times a week
to do this “manually” … an hour ago I was about to do this again when I
realized with a little guilt that not everyone on my team is
sufficiently proficient with Clojure REPLs to do this themselves, and
that the right thing to do there, for their sake and mine, would be to
add this to dad as a new feature.

This has the added benefit in that it spurred me to finally start using
a proper options-parsing library to validate options, generate help
text, etc (in the prior commit).

(I was about to use tools.cli for that when I noticed a mention of
cli-matic in the tools.cli readme, which was great because cli-matic
seems excellent!)