TE: switch a dep to a fork
Because that fork has less dependencies — it just removes some deps that
don’t actually seem necessary. I’m going to try to submit the change
upstream and if it’s accepted, I’ll then switch this project back to the
upstream project and delete the fork.

This has various benefits, including but not limited to:

* Reducing the time required to compile a native-image for MacOS, on my
  laptop, from ~150 to ~104 secs
* Reducing the memory required to compile a native-image for MacOS, on
  my laptop.
  * I’ve been struggling to compile with 6GB of RAM allocated to
    native-image, and I’ve been trying to make that work because GitHub
    Actions virtual machines are limited to 7GB of RAM, and it’s super
    useful to be able to build and compile executables as part of our
    automation pipelines.
* Reducing the sizes of distribution artifacts (uberjar & native image)