Custom tag exec should throw exceptions on error
In a downstream project I’ve been debugging a template that uses this
tag, and it’s failing, and I’ve realized that the current approach to
error handling — including the error output in the generated document
and continuing — is terrible. It’s basically failing silently.

So this changes the tag to fail fast on error by throwing Exceptions, which is similar to how Selmer’s built-in tags work.