Re-render docs
Apparently I forgot to re-render them at some point, so the built docs
have been out of sync with the templates+data. I discovered this
randomly by attempting to follow a link from the root readme to
doc/built/docs-as-data.md and of course that link was bad; it returned
a 404 not found error page.

To prevent this from recurring, I’ve created FundingCircle/DaD#32:
“An automation to ensure that the built docs are in sync with the
templates+data”